Bug Fix: Move cancel call to downloadSub after client.close()
Because canceling a stream subscription is an async method, and we weren't awaiting terminateConnection in onByteExceeded and onByteExactMatch, connections would start fixing the temp files when the cancel event wasn't yet completed and therefore, the connections were still receiving data, leading to a spam of completionSignals to the engine, connections constantly exceeding their boundaries, and various other unpredictable outcomes. Closing the client first ensures that the connection termination takes effect instantly due to it not being async.

This is a major bug but is not included in any release. It was introduced in commit 5b030ca975930de29a623d088a120919d772049e and caught during development.